+++ src/app/convenience-file-types.test.js
@@ -1,13 +1,49 @@
 import expect from 'expect';
 import {
+  expandGenesJson,
   expandCellsJson,
   expandAnndataZarr,
   expandAnndataCellsZarr,
   expandAnndataCellSetsZarr,
   expandAnndataExpressionMatrixZarr,
+  expandClustersJson,
 } from './convenience-file-types';
 
 describe('src/app/convenience-file-types.js', () => {
+  describe('expandClustersJson', () => {
+    it('expands clusters.json', () => {
+      expect(expandClustersJson({
+        fileType: 'clusters.json',
+        url: 'http://localhost:8000/clusters.json',
+      })).toEqual([
+        {
+          fileType: 'obsFeatureMatrix.clusters.json',
+          url: 'http://localhost:8000/clusters.json',
+          coordinationValues: {
+            obsType: 'cell',
+            featureType: 'gene',
+          },
+        },
+      ]);
+    });
+  });
+  describe('expandGenesJson', () => {
+    it('expands', () => {
+      expect(expandGenesJson({
+        fileType: 'genes.json',
+        url: 'http://localhost:8000/genes.json',
+      })).toEqual([
+        {
+          fileType: 'obsFeatureMatrix.genes.json',
+          url: 'http://localhost:8000/genes.json',
+          coordinationValues: {
+            obsType: 'cell',
+            featureType: 'gene',
+          },
+        },
+      ]);
+    });
+  });
   describe('expandCellsJson', () => {
     it('expands when there are no options', () => {
       expect(expandCellsJson({