+++ src/loaders/types.js
@@ -5,11 +5,11 @@ import {
 
 import JsonLoader from './JsonLoader';
 import MatrixZarrLoader from './MatrixZarrLoader';
-import GenesJsonAsMatrixZarrLoader from './GenesJsonAsMatrixZarrLoader';
-import ClustersJsonAsMatrixZarrLoader from './ClustersJsonAsMatrixZarrLoader';
+import ClustersJsonAsMatrixZarrLoader from './clusters-json-loaders/ClustersJsonAsMatrixZarrLoader';
+import GenesJsonAsMatrixZarrLoader from './genes-json-loaders/GenesJsonAsMatrixZarrLoader';
 import RasterJsonLoader from './RasterJsonLoader';
 import OmeZarrLoader from './OmeZarrLoader';
-import CellSetsJsonLoader from './CellSetsJsonLoader';
+import ObsSetsJsonLoader from './ObsSetsJsonLoader';
 import AnnDataLoaders from './anndata-loaders';
 import GenomicProfilesZarrLoader from './GenomicProfilesZarrLoader';
 import { AnnDataSource, ZarrDataSource, JsonSource } from './data-sources';
@@ -17,6 +17,8 @@ import CellsJsonAsObsLabelsLoader from './cells-json-loaders/CellsJsonAsObsLabel
 import CellsJsonAsObsEmbeddingLoader from './cells-json-loaders/CellsJsonAsObsEmbedding';
 import CellsJsonAsObsLocationsLoader from './cells-json-loaders/CellsJsonAsObsLocations';
 import CellsJsonAsObsSegmentationsLoader from './cells-json-loaders/CellsJsonAsObsSegmentations';
+import ClustersJsonAsObsFeatureMatrixLoader from './clusters-json-loaders/ClustersJsonAsObsFeatureMatrix';
+import GenesJsonAsObsFeatureMatrixLoader from './genes-json-loaders/GenesJsonAsObsFeatureMatrix';
 
 export const fileTypeToLoaderAndSource = {
   [FileType.OBS_FEATURE_MATRIX_ANNDATA_ZARR]: [
@@ -34,6 +36,10 @@ export const fileTypeToLoaderAndSource = {
   [FileType.OBS_SEGMENTATIONS_ANNDATA_ZARR]: [
     AnnDataSource, AnnDataLoaders.ObsSegmentationsAnndataLoader,
   ],
+  [FileType.OBS_SETS_JSON]: [JsonSource, ObsSetsJsonLoader],
+  [FileType.OBS_SETS_CELL_SETS_JSON]: [JsonSource, ObsSetsJsonLoader],
+  [FileType.OBS_FEATURE_MATRIX_CLUSTERS_JSON]: [JsonSource, ClustersJsonAsObsFeatureMatrixLoader],
+  [FileType.OBS_FEATURE_MATRIX_GENES_JSON]: [JsonSource, GenesJsonAsObsFeatureMatrixLoader],
   [FileType.OBS_LABELS_CELLS_JSON]: [JsonSource, CellsJsonAsObsLabelsLoader],
   [FileType.OBS_EMBEDDING_CELLS_JSON]: [JsonSource, CellsJsonAsObsEmbeddingLoader],
   [FileType.OBS_LOCATIONS_CELLS_JSON]: [JsonSource, CellsJsonAsObsLocationsLoader],
@@ -46,7 +52,7 @@ export const fileTypeToLoaderAndSource = {
   [FileType.NEIGHBORHOODS_JSON]: [JsonSource, JsonLoader],
   [FileType.RASTER_JSON]: [JsonSource, RasterJsonLoader],
   [FileType.RASTER_OME_ZARR]: [ZarrDataSource, OmeZarrLoader],
-  [FileType.CELL_SETS_JSON]: [JsonSource, CellSetsJsonLoader],
+  [FileType.CELL_SETS_JSON]: [JsonSource, ObsSetsJsonLoader],
   [FileType.ANNDATA_CELL_SETS_ZARR]: [AnnDataSource, AnnDataLoaders.CellSetsZarrLoader],
   [FileType.ANNDATA_CELLS_ZARR]: [AnnDataSource, AnnDataLoaders.CellsZarrLoader],
   [FileType.ANNDATA_EXPRESSION_MATRIX_ZARR]: [AnnDataSource, AnnDataLoaders.MatrixZarrLoader],